atualizado

+++ app/api/post.py
@@ -3,11 +3,11 @@ from datetime import datetime
 from flask import request, Flask, url_for
 from werkzeug.utils import secure_filename
 from flask_restful import Resource, reqparse
-from flask_jwt_extended import jwt_required, get_jwt_identity
+from flask_jwt_extended import jwt_required, get_jwt_identity, get_jwt
 from app.extensions import db
-from app.models import Usuario, Foto, Post
+from app.models import Usuario, Foto, Post, TokenBlacklist
 
-# Configurações para o upload de imagens de perfil
+
 UPLOAD_FOLDER_PERFIL = 'uploads/fotos_perfil'
 UPLOAD_FOLDER = 'uploads/fotos_posts'
 ALLOWED_EXTENSIONS = {'png', 'jpg', 'jpeg', 'gif'}
@@ -25,35 +25,35 @@ def allowed_file(filename):
 
 class PostListResource(Resource):
     def get(self):
-        # Debug: Verifica se a função está sendo chamada
+        
         print("Entrou na função get()")
 
         parser = reqparse.RequestParser()
         parser.add_argument('page', type=int, default=1,
-                            location='args')  # Parâmetro de consulta
+                            location='args')  
         parser.add_argument('limit', type=int, default=10,
-                            location='args')  # Parâmetro de consulta
+                            location='args')  
         parser.add_argument('user_id', type=int, required=False,
-                            location='args')  # Parâmetro de consulta
-        args = parser.parse_args()  # Extrai os parâmetros da requisição
+                            location='args')  
+        args = parser.parse_args()  
 
         query = Post.query
         if args['user_id']:
-            # Filtra posts por ID do usuário
+            
             query = query.filter_by(id_usuario=args['user_id'])
 
-        # Paginação dos posts
+        
         posts = query.order_by(Post.data_criacao.desc()).paginate(
             page=args['page'], per_page=args['limit'], error_out=False)
 
-        # Formata os dados para incluir a URL da imagem, o ID do usuário, a URL da foto de perfil e outras informações
+       
         fotos = []
         for post in posts.items:
             usuario = Usuario.query.get(post.id_usuario)
             imagem_url = url_for(
                 'upload.uploaded_file', filename=post.imagem.split('\\')[-1], _external=True)
 
-            # Busca a foto de perfil do usuário
+            
             foto_perfil = Foto.query.filter_by(id_usuario=usuario.id).first()
             foto_perfil_url = None
             if foto_perfil:
@@ -61,48 +61,48 @@ class PostListResource(Resource):
                     'upload.uploaded_file_perfil', filename=foto_perfil.imagem.split('\\')[-1], _external=True)
 
             fotos.append({
-                "id": post.id,  # ID do post
-                "imagem_url": imagem_url,  # URL completa da imagem do post
-                "data_criacao": post.data_criacao.isoformat(),  # Data de criação formatada
-                "legenda": post.legenda,  # Legenda do post
+                "id": post.id, 
+                "imagem_url": imagem_url,  
+                "data_criacao": post.data_criacao.isoformat(),  
+                "legenda": post.legenda,  
                 "usuario": {
-                    "id": usuario.id,  # ID do usuário
-                    "username": usuario.username,  # Nome de usuário
-                    "foto_perfil_url": foto_perfil_url  # URL da foto de perfil do usuário
+                    "id": usuario.id,  
+                    "username": usuario.username,  
+                    "foto_perfil_url": foto_perfil_url  
                 }
             })
 
         return {
-            "fotos": fotos,  # Lista de posts formatados
-            "total": posts.total,  # Total de posts
-            "page": posts.page,  # Página atual
-            "pages": posts.pages  # Total de páginas
+            "fotos": fotos,  
+            "total": posts.total,  
+            "page": posts.page,  
+            "pages": posts.pages  
         }, 200
 
 
 class PostResource(Resource):
     def get(self, post_id):
-        # Busca o post pelo ID
+        
         post = Post.query.get(post_id)
         if not post:
             return {"error": "Post não encontrado"}, 404
 
-        # Busca o usuário associado ao post
+        
         usuario = Usuario.query.get(post.id_usuario)
         if not usuario:
             return {"error": "Usuário associado ao post não encontrado"}, 404
 
-        # Gera a URL da imagem
+        
         imagem_url = url_for(
             'upload.uploaded_file', filename=post.imagem.split('\\')[-1], _external=True)
-     # Busca a foto de perfil do usuário
+     
         foto_perfil = Foto.query.filter_by(id_usuario=usuario.id).first()
         foto_perfil_url = None
         if foto_perfil:
             foto_perfil_url = url_for(
                 'upload.uploaded_file_perfil', filename=foto_perfil.imagem.split('\\')[-1], _external=True)
 
-        # Retorna os dados do post formatados
+        
         return {
             "id": post.id,
             "imagem_url": imagem_url,
@@ -119,6 +119,9 @@ class PostResource(Resource):
 class CreatePostResource(Resource):
     @jwt_required()
     def post(self):
+        blacklist_check = check_token_blacklist()
+        if blacklist_check:
+            return blacklist_check
         user_id = get_jwt_identity()
         usuario = Usuario.query.get(user_id)
 
@@ -134,31 +137,31 @@ class CreatePostResource(Resource):
             return {"error": "Nome do arquivo inválido"}, 400
 
         if file and allowed_file(file.filename):
-            # Obtém a extensão do arquivo
+            
             ext = file.filename.rsplit('.', 1)[1].lower()
 
-            # Obtém a legenda
+           
             legenda = request.form.get('legenda')
             if not legenda:
                 return {"error": "Legenda é obrigatória"}, 400
 
-            # Cria um novo post com um valor temporário para imagem
+            
             novo_post = Post(
                 legenda=legenda,
-                imagem="temp.jpg",  # Valor temporário
+                imagem="temp.jpg",  
                 id_usuario=user_id,
                 data_criacao=datetime.now()
             )
 
             db.session.add(novo_post)
-            db.session.commit()  # Agora novo_post.id existe
+            db.session.commit()  
 
-            # Define o nome final do arquivo
+            
             filename = f"{user_id}_{novo_post.id}.{ext}"
             filepath = os.path.join(UPLOAD_FOLDER, filename)
             file.save(filepath)
 
-            # Atualiza o caminho correto da imagem
+            
             novo_post.imagem = filepath
             db.session.commit()
 
@@ -177,39 +180,45 @@ class CreatePostResource(Resource):
 
 
 class DeletePostResource(Resource):
-    @jwt_required()  # Requer autenticação JWT
+    @jwt_required()  
     def delete(self, post_id):
-        # Obtém o ID do usuário logado a partir do token JWT
+        blacklist_check = check_token_blacklist()
+        if blacklist_check:
+            return blacklist_check
+        
         user_id = int(get_jwt_identity())
 
-        # Busca o post pelo ID
+       
         post = Post.query.get(post_id)
         if not post:
             return {"error": "Post não encontrado"}, 404
 
-        # Verifica se o usuário logado é o proprietário do post
+       
         if post.id_usuario != user_id:
             return {"error": "Você não tem permissão para excluir este post"}, 403
 
         try:
-            # Remove o arquivo de imagem associado ao post
+            
             if os.path.exists(post.imagem):
                 os.remove(post.imagem)
 
-            # Remove o post do banco de dados
+            
             db.session.delete(post)
             db.session.commit()
 
             return {"message": "Post excluído com sucesso"}, 200
         except Exception as e:
-            db.session.rollback()  # Desfaz a transação em caso de erro
+            db.session.rollback() 
             return {"error": f"Erro ao excluir o post: {str(e)}"}, 500
 
 
 class UploadFotoPerfilResource(Resource):
     @jwt_required()
     def post(self):
-        # Obtém o ID do usuário logado a partir do token JWT
+        blacklist_check = check_token_blacklist()
+        if blacklist_check:
+            return blacklist_check
+        
         user_id = get_jwt_identity()
         usuario = Usuario.query.get(user_id)
         print("user_id, usuario:  ", user_id, usuario)
@@ -225,24 +234,24 @@ class UploadFotoPerfilResource(Resource):
             return {"error": "Nome do arquivo inválido"}, 400
 
         if file and allowed_file(file.filename):
-            # Obtém a extensão do arquivo
+           
             ext = file.filename.rsplit('.', 1)[1].lower()
 
-            # Define o nome final do arquivo
+           
             filename = f"{user_id}_perfil.{ext}"
             filepath = os.path.join(UPLOAD_FOLDER_PERFIL, filename)
 
-            # Salva o arquivo no sistema de arquivos
+            
             file.save(filepath)
 
-            # Cria ou atualiza a foto de perfil do usuário
+            
             foto_perfil = Foto.query.filter_by(id_usuario=user_id).first()
             if foto_perfil:
-                # Atualiza a foto de perfil existente
+                
                 foto_perfil.imagem = filepath
                 foto_perfil.data_criacao = datetime.utcnow()
             else:
-                # Cria uma nova foto de perfil
+                
                 foto_perfil = Foto(
                     imagem=filepath,
                     id_usuario=user_id,
@@ -267,24 +276,30 @@ class UploadFotoPerfilResource(Resource):
 
 class FotoPerfilResource(Resource):
     def get(self, usuario_id):
-        """
-        Retorna a URL da foto de perfil do usuário com base no ID fornecido.
-        """
-        # Busca o usuário no banco de dados
+        
         usuario = Usuario.query.get(usuario_id)
         if not usuario:
             return {"error": "Usuário não encontrado"}, 404
 
-        # Busca a foto de perfil do usuário
+        
         foto_perfil = Foto.query.filter_by(id_usuario=usuario_id).first()
         if not foto_perfil:
             return {"error": "Foto de perfil não encontrada"}, 404
 
-        # Gera a URL da imagem de perfil
+        
         foto_perfil_url = url_for(
             'upload.uploaded_file_perfil', filename=foto_perfil.imagem.split('\\')[-1], _external=True)
 
-        # Retorna a URL da foto de perfil
+        
         return {
             "foto_perfil_url": foto_perfil_url
         }, 200
+
+
+
+def check_token_blacklist():
+    decoded_token = get_jwt()
+    jti = decoded_token["jti"]
+    if TokenBlacklist.query.filter_by(token=jti).first():
+        return {"error": "Token inválido. Faça login novamente."}, 401
+    return None
